Scaffolding cleanup
