+++ packages/devtools_app/lib/src/app.dart
@@ -184,7 +184,7 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
     return MaterialPage(
       child: DevToolsScaffold.withChild(
         key: const Key('not-found'),
-        ideTheme: ideTheme,
+        embed: isEmbedded(args),
         child: CenteredMessage("'$page' not found."),
       ),
     );
@@ -197,12 +197,13 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
     DevToolsNavigationState? __,
   ) {
     final vmServiceUri = params['uri'];
+    final embed = isEmbedded(params);
 
     // Always return the landing screen if there's no VM service URI.
     if (vmServiceUri?.isEmpty ?? true) {
       return DevToolsScaffold.withChild(
         key: const Key('landing'),
-        ideTheme: ideTheme,
+        embed: embed,
         child: LandingScreenBody(sampleData: widget.sampleData),
       );
     }
@@ -213,7 +214,6 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
     if (page?.isEmpty ?? true) {
       page = params['page'];
     }
-    final embed = params['embed'] == 'true';
     final hide = {...?params['hide']?.split(',')};
     return Initializer(
       url: vmServiceUri,
@@ -247,7 +247,7 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
             );
           },
           child: DevToolsScaffold.withChild(
-            ideTheme: ideTheme,
+            embed: embed,
             child: CenteredMessage(
               page != null
                   ? 'The "$page" screen is not available for this application.'
@@ -262,24 +262,26 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
   /// The pages that the app exposes.
   Map<String, UrlParametersBuilder> get pages {
     return _routes ??= {
-      homePageId: _buildTabbedPage,
+      homeScreenId: _buildTabbedPage,
       for (final screen in widget.screens)
         screen.screen.screenId: _buildTabbedPage,
-      snapshotPageId: (_, __, args, ___) {
+      snapshotScreenId: (_, __, args, ___) {
         final snapshotArgs = OfflineDataArguments.fromArgs(args);
+        final embed = isEmbedded(args);
         return DevToolsScaffold.withChild(
           key: UniqueKey(),
-          ideTheme: ideTheme,
+          embed: embed,
           child: MultiProvider(
             providers: _providedControllers(offline: true),
             child: OfflineScreenBody(snapshotArgs, _screens),
           ),
         );
       },
-      appSizePageId: (_, __, ___, ____) {
+      appSizeScreenId: (_, __, args, ____) {
+        final embed = isEmbedded(args);
         return DevToolsScaffold.withChild(
           key: const Key('appsize'),
-          ideTheme: ideTheme,
+          embed: embed,
           child: MultiProvider(
             providers: _providedControllers(),
             child: const AppSizeBody(),
@@ -288,6 +290,8 @@ class DevToolsAppState extends State<DevToolsApp> with AutoDisposeMixin {
       },
     };
   }
+
+  bool isEmbedded(Map<String, String?> args) => args['embed'] == 'true';
 
   Map<String, UrlParametersBuilder>? _routes;
